Update feet to meters converter.cpp

+++ feet to meters converter.cpp
@@ -4,8 +4,8 @@ using namespace std;
 int main()
 {
 	int feet;
-	int meters;
-	int converter = 0.3;
+	double meters;
+	double converter = 0.3;
 
 	cout << "Enter height in feet: ";
 	cin >> feet;
